Add default-light.vim
Vim's default color scheme has different colors for dark background.
Default2 does not. This means we can't load the default scheme from
$VIMRUNTIME/colors and extend it, but we have to include the light
version of the default theme.

default-light.vim is the result of :highlight when running Vim with the
default color scheme on a light background without any configuration,
except for `:syntax on`.

The linked and cleared highlights and all gui rules were removed. After
that, it was turned into a color scheme by removing the x-es and adding
`highlight` before every line.